[Router] extract shared Anthropic response-shape helpers from client.go
ToOpenAIResponseBody is the inverse of the Anthropic outbound emitter
that PR4 is building. The field-mapping tables (Anthropic content
block → OpenAI message; Anthropic stop_reason → OpenAI finish_reason;
Anthropic usage → OpenAI usage) are the same tables read in reverse.
Extracting them into small reusable helpers lets the symmetric outbound
emitter consume them without duplicating logic.

Helpers extracted:

- openAIFinishReasonFromAnthropic: stop_reason → finish_reason mapping
  (was inline switch); also collapses Anthropic-only "pause_turn" and
  "refusal" to "stop" because the OpenAI envelope cannot represent
  them.
- captureAnthropicStopReasonIntoExt: records Anthropic-only stop
  reasons and the stop_sequence string onto IRExtensions when a side
  channel is wired, so the symmetric outbound emitter can restore
  them.
- extractThinkingBlocks: records ThinkingBlock signatures onto
  IRExtensions for the multi-turn extended-thinking round-trip.
- extractAnthropicUsageIntoExt: copies cache_*_input_tokens, the
  CacheCreation TTL breakdown, and server_tool_use counts onto
  IRExtensions so the symmetric emitter can replay them on response.

Behavior of the legacy entrypoint (ToOpenAIResponseBody, called with
no side channel) is preserved byte-for-byte: each helper is a no-op
when ext is nil. New refactor-parity tests cover the helpers
individually and assert the legacy entrypoint still produces a
byte-identical OpenAI envelope for an Anthropic response that
contains thinking, refusal stop_reason, and cache usage.
